fix reply conent with InnerHtml
